Moved Base16 to util package
