+++ src/test/java/net/tassia/pancake/Base16Test.kt
@@ -1,5 +1,6 @@
 package net.tassia.pancake
 
+import net.tassia.pancake.util.Base16
 import kotlin.random.Random
 import kotlin.test.Test
 import kotlin.test.assertTrue
